{{cookiecutter.github_repository_name}}/src/{{cookiecutter.app_name}}/users/serializers.py
from rest_framework import serializers

from {{cookiecutter.app_name}}.users.models import User


class UserSerializer(serializers.ModelSerializer):

    class Meta:
        model = User
        fields = ('username', 'first_name', 'last_name', 'email')
        write_only_fields = ('password',)

    def create(self, validated_data):
        # call create_user on user object. Without this
        # the password will be stored in plain text.
        user = User.objects.create_user(**validated_data)
        return user

