{{cookiecutter.github_repository_name}}/src/requirements.txt
# Django
Django==1.9.7

# Environmental var configuration
django-configurations==1.0

# For the persistence stores
psycopg2==2.6.1
dj-database-url==0.4.0

# Models
django-model-utils==2.4

# Rest apis
djangorestframework==3.3.3
Markdown==2.6.6
django-filter==0.13.0

# Static files
whitenoise==3.0

# Media files
boto==2.39.0
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.3

# For asynchronous queuing
django-rq==0.9.0

# Time zones support
pytz==2016.3

# Admion Theme
django-suit==0.2.18

# WSGI Handler
gevent==1.0.2
gunicorn==19.3.0

# Static and Media Storage
django-storages-redux==1.2.3

# Caching
hiredis==0.2.0
django-redis-cache==1.6.5

# Django-secure
django-secure==1.0.1

# App Monitoring
newrelic==2.50.0.39

# Debugging
ipdb==0.9.0
ipython==4.1.2

# Docs
mkdocs==0.15.3

# Command Utils
Fabric==1.10.2

# Mocking
mock==1.3.0

pytest==2.9.1
pytest-django==2.9.1

# Code Coverage
coverage==4.0.3

# Style Enforcement
flake8==2.5.4
